refactor: auth 도메인 -UserInfoDto 클래스 패키지 이동
해당 클래스 -UserInfo 클래스 내부에서만 사용되기 때문에
*/auth/info/dto 패키지로 옮김
